test(vscode): isolate markdown annotation mutation helper
