+++ packages/kilo-vscode/tests/unit/markdown-annotation-layer.test.ts
@@ -1,5 +1,5 @@
 import { describe, expect, it } from "bun:test"
-import { isAnnotationMutation } from "../../webview-ui/agent-manager/MarkdownAnnotationLayer"
+import { isAnnotationMutation } from "../../webview-ui/agent-manager/markdown-annotation-mutation"
 
 function mutation(target: Node): Pick<MutationRecord, "target"> {
   return { target }
